teamf-backend/pyproject.toml
[tool.poetry]
name = "teamfbackend"
version = "0.1.0"
description = ""
authors = ["Matheus Felizardo <matheus.felizardo2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
annotated-types = "^0.6.0"
blinker = "^1.7.0"
click = "^8.1.7"
dnspython = "^2.4.2"
flask = "^3.0.0"
flask-cors = "^4.0.0"
flask-jwt-extended = "^4.5.3"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
markupsafe = "^2.1.3"
pyjwt = "^2.8.0"
pymongo = "^4.6.0"
python-dotenv = "^1.0.0"
simple-chalk = "^0.1.0"
typing-extensions = "^4.8.0"
werkzeug = "^3.0.1"
pydantic-core = "2.10.1"
pydantic = "2.4.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
